Upgrade func store_image, now without duplicats.

+++ backend/tools/image_validator.py
@@ -2,33 +2,35 @@ import imghdr
 import uuid
 import os
 import base64
-from time import time
+from hashlib import md5
+
+
+async def decoCut_base64(str_base64):
+    return base64.b64decode(str_base64[str_base64.index(",") + 1:])
 
 
 async def is_image(str_base64):
-    flag = False
+    # possible optimisation -- module tempfile
     if str_base64 == []:
-        return flag
-    cuted_base64 = base64.b64decode(str_base64[str_base64.index(",") + 1:])
+        return None
     path = f"/tmp/{uuid.uuid1()}"
+    binary = await decoCut_base64(str_base64)
     with open(path, "wb") as tmpfile:
-        tmpfile.write(cuted_base64)
-        extension = imghdr.what(path)
-        if extension:
-            flag = True
+        tmpfile.write(binary)
+    extension = imghdr.what(path)
     os.remove(path)
-    return flag, extension
+    return extension
 
 
 async def store_image(str_base64, extension, chat):
     local_path = f'./../../public/images/{chat}/'
     if not os.path.isdir(local_path):
         os.mkdir(local_path)  # descriptors
-    binary = base64.b64decode(str_base64[str_base64.index(",") + 1:])
-    name_file = str(int(time()))
+    binary = await decoCut_base64(str_base64)
+    name_file = md5(binary).hexdigest()
     full_local_path = local_path + name_file + "." + extension
-    with open(full_local_path, 'wb') as image:
-        image.write(binary)
-    # TODO: send full path to image
+    if not os.path.isfile(full_local_path):
+        with open(full_local_path, 'wb') as image:
+            image.write(binary)
     path = '/images/' + chat + "/" + name_file + "." + extension
     return path
